capture: state the Graph provider guarantees where callers read them
Craftsmanship pass on the two refusal guards.

The non-empty check at the backfill call site is gone: ListAfter now refuses a
listed message that names no folder, so the conjunct could never be false, and
leaving it in advertised a doubt about the boundary that the guard itself
retired. The guarantee moves onto MessageRef's doc, where a caller reads it.

The extracted transport file claimed the OAuth handshake sits on it. It does
not — NewOAuth delegates to capture/oauthflow, which has its own client and its
own classification — so the split's stated reason is now the true one.

Both degraded-response tests move to client_test.go, where the httptest-backed
client tests live, and onto the shared JSON stub; one of them was named for a
backfill behaviour it never exercised (that half is already covered by
TestBackfillStopsWhenTheSentFolderCannotBeResolved).

The migration's residual list said three and enumerated four.
