+++ backend/internal/modules/capture/graph/backfill_test.go
@@ -6,8 +6,6 @@ package graph
 import (
 	"context"
 	"errors"
-	"net/http"
-	"net/http/httptest"
 	"testing"
 	"time"
 
@@ -153,42 +151,5 @@ func TestBackfillStopsWhenTheSentFolderCannotBeResolved(t *testing.T) {
 	}
 	if len(sink.recs) != 0 {
 		t.Fatalf("%d records captured, want none — nothing may land with guessed provenance", len(sink.recs))
-	}
-}
-
-// A 2xx that names no folder is not an answer. Left as "", it would compare
-// equal to the empty parentFolderId of any message the same degraded response
-// returned, so the ABSENCE of the evidence would read as the evidence and
-// attest the whole page — and the activity natural key would make it permanent.
-func TestBackfillStopsWhenTheSentFolderIDComesBackEmpty(t *testing.T) {
-	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		//craft:ignore swallowed-errors test stub write; a short write surfaces as the client-side assertion failure
-		_, _ = w.Write([]byte(`{}`))
-	}))
-	defer srv.Close()
-
-	_, err := NewAPI(srv.Client(), srv.URL).SentFolderID(context.Background(), "access-2")
-	if !errors.Is(err, ErrUnreachable) {
-		t.Fatalf("SentFolderID on an id-less 2xx = %v, want ErrUnreachable — an empty id must never reach the comparison", err)
-	}
-}
-
-// The mirror of the empty-folder-id case, and it must fail the same way. A
-// listed message naming no parent folder cannot be captured on a guess in
-// either direction: attested, an empty id would match an unresolved folder;
-// un-attested, the activity natural key would permanently discard evidence the
-// owner really did produce. The page stops instead.
-func TestListAfterRefusesAMessageWithNoParentFolder(t *testing.T) {
-	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		//craft:ignore swallowed-errors test stub write; a short write surfaces as the client-side assertion failure
-		_, _ = w.Write([]byte(`{"value":[{"id":"m-truncated"}]}`))
-	}))
-	defer srv.Close()
-
-	_, _, err := NewAPI(srv.Client(), srv.URL).ListAfter(context.Background(), "access-2", time.Time{}, "", 100)
-	if !errors.Is(err, ErrUnreachable) {
-		t.Fatalf("ListAfter on a folder-less message = %v, want ErrUnreachable", err)
 	}
 }
